requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
google-api-core==1.25.1
google-auth==1.25.0
google-cloud-speech==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
idna==2.10
itsdangerous @ file:///home/ktietz/src/ci/itsdangerous_1611932585308/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
libcst==0.3.16
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1607027305082/work
mypy-extensions==0.4.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1607448156598/work
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
scipy @ file:///tmp/build/80754af9/scipy_1612469580794/work
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
